// migrations/snapshots/8ee01b02869c34d67e06edae37083fddfd3b11b8f954e4e3ed8f96d8032164c9/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Loger: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly process: 'CREATE' | 'UPDATE' | 'DELETE' | 'RESTORE' | 'LOGIN' | 'LOGOUT';
      readonly detail: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly MobilPushSubscription: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Notification: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly _type: CodecTypes['pg/text@1']['output'];
      readonly refId: CodecTypes['pg/text@1']['output'] | null;
      readonly message: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly NotificationUser: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly notificationId: CodecTypes['pg/int4@1']['output'];
      readonly readStatus: CodecTypes['pg/bool@1']['output'];
      readonly readAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly deliveredAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly Otp: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly purpose: 'LOGIN' | 'RESET_PASSWORD' | 'VERIFY_EMAIL';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Permissions: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
    };
    readonly Role: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly roleType: 'OTHER' | 'SUPER_ADMIN';
    };
    readonly RolePermission: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly permissionId: CodecTypes['pg/int4@1']['output'];
      readonly canRead: CodecTypes['pg/bool@1']['output'];
      readonly canWrite: CodecTypes['pg/bool@1']['output'];
      readonly canUpdate: CodecTypes['pg/bool@1']['output'];
      readonly canDelete: CodecTypes['pg/bool@1']['output'];
      readonly canRestore: CodecTypes['pg/bool@1']['output'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly registeredViaGoogle: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly WebPushSubscription: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly endpoint: CodecTypes['pg/text@1']['output'];
      readonly p256dh: CodecTypes['pg/text@1']['output'];
      readonly auth: CodecTypes['pg/text@1']['output'];
      readonly expirationTime: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Loger: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly process: 'CREATE' | 'UPDATE' | 'DELETE' | 'RESTORE' | 'LOGIN' | 'LOGOUT';
      readonly detail: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly MobilPushSubscription: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Notification: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly _type: CodecTypes['pg/text@1']['input'];
      readonly refId: CodecTypes['pg/text@1']['input'] | null;
      readonly message: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly NotificationUser: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly notificationId: CodecTypes['pg/int4@1']['input'];
      readonly readStatus: CodecTypes['pg/bool@1']['input'];
      readonly readAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly deliveredAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly Otp: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly purpose: 'LOGIN' | 'RESET_PASSWORD' | 'VERIFY_EMAIL';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Permissions: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
    };
    readonly Role: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly roleType: 'OTHER' | 'SUPER_ADMIN';
    };
    readonly RolePermission: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly roleId: CodecTypes['pg/int4@1']['input'];
      readonly permissionId: CodecTypes['pg/int4@1']['input'];
      readonly canRead: CodecTypes['pg/bool@1']['input'];
      readonly canWrite: CodecTypes['pg/bool@1']['input'];
      readonly canUpdate: CodecTypes['pg/bool@1']['input'];
      readonly canDelete: CodecTypes['pg/bool@1']['input'];
      readonly canRestore: CodecTypes['pg/bool@1']['input'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly roleId: CodecTypes['pg/int4@1']['input'];
      readonly registeredViaGoogle: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly WebPushSubscription: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly endpoint: CodecTypes['pg/text@1']['input'];
      readonly p256dh: CodecTypes['pg/text@1']['input'];
      readonly auth: CodecTypes['pg/text@1']['input'];
      readonly expirationTime: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;